+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.6.3]() - 2023-
+### Fixed
+- Wait-OnJob MAX_RETRIES calculation is set based on the SLEEP_INTERVAL
+
 ### Changed
 - Update-OMEConfiguration added -UpdateSchedule parameter with the option to "StageForNextReboot" or "RebootNow"